+++ src/main/webapp/wise5/directives/component/component.js
@@ -84,7 +84,7 @@ var ComponentController = function ComponentController($injector, $scope, $compi
         $scope.nodeAuthoringController = $scope.$parent.nodeAuthoringController;
     }
 
-    if ($scope.mode === 'authoring' && (componentContent.type == 'HTML' || componentContent.type == 'ConceptMap' || componentContent.type == 'Animation' || componentContent.type == 'Graph')) {
+    if ($scope.mode === 'authoring' && (componentContent.type == 'HTML' || componentContent.type == 'ConceptMap' || componentContent.type == 'Animation' || componentContent.type == 'Graph' || componentContent.type == 'Label')) {
         $scope.componentTemplatePath = this.NodeService.getComponentAuthoringTemplatePath(componentContent.type);
     } else {
         $scope.componentTemplatePath = this.NodeService.getComponentTemplatePath(componentContent.type);
